+++ frontend/app/app/(pages)/auth/sign_in/page.tsx
@@ -4,9 +4,9 @@ import NextLink from "next/link";
 
 import { Center, Divider, Heading, Link, Text } from "@chakra-ui/react";
 
+import WithRedirectIfLoggedIn from "@/app/features/auth/accessControl/WithRedirectIfLoggedIn";
 import GuestLoginButton from "@/app/features/auth/guestLogin/GuestLoginButton";
 import SignInForm from "@/app/features/auth/signIn/SignInForm";
-import WithRedirectIfLoggedIn from "@/app/components/HOCs/WithRedirectIfLoggedIn";
 import RoundedCenteredBox from "@/app/components/styledWrappers/RoundedCenteredBox";
 
 const SignInPage: React.FC = () => {
