+++ frontend/app/app/(pages)/users/comments/page.tsx
@@ -14,8 +14,8 @@ import {
   Text
 } from "@chakra-ui/react";
 
+import WithRedirectIfLoggedOut from "@/app/features/auth/accessControl/WithRedirectIfLoggedOut";
 import { useGetLoginState } from "@/app/hooks/recoil/loginState/useGetLoginState";
-import WithRedirectIfLoggedOut from "@/app/components/HOCs/WithRedirectIfLoggedOut";
 import GoBackLink from "@/app/components/atoms/GoBackLink";
 import UpdatedDateText from "@/app/components/atoms/texts/UpdatedDateText";
 import PostCommentTags from "@/app/components/molecules/posts/comments/PostCommentTags";
