+++ frontend/app/app/(pages)/users/likes/page.tsx
@@ -2,8 +2,8 @@
 
 import { Text } from "@chakra-ui/react";
 
+import WithRedirectIfLoggedOut from "@/app/features/auth/accessControl/WithRedirectIfLoggedOut";
 import { useGetLikedPortfolios } from "@/app/hooks/swr/me/useGetLikedPortfolios";
-import WithRedirectIfLoggedOut from "@/app/components/HOCs/WithRedirectIfLoggedOut";
 import GoBackLink from "@/app/components/atoms/GoBackLink";
 import LoadingSpinner from "@/app/components/atoms/LoadingSpinner";
 import PostCard from "@/app/components/organisms/posts/PostCard";
