+++ frontend/app/app/features/auth/accessControl/__tests__/WithRedirectIfLoggedIn.test.tsx
@@ -3,12 +3,13 @@ import { render } from "@testing-library/react";
 import { validUserData } from "@/__tests__/fixtures/auth/validUserData";
 import { mockNavigation, replaceMock } from "@/__tests__/mocks/mockNavigation";
 import { useGetLoginState } from "@/app/hooks/recoil/loginState/useGetLoginState";
-import WithRedirectIfLoggedOut from "@/app/components/HOCs/WithRedirectIfLoggedOut";
+
+import WithRedirectIfLoggedIn from "../WithRedirectIfLoggedIn";
 
 jest.mock("next/navigation", () => mockNavigation);
 jest.mock("@/app/hooks/recoil/loginState/useGetLoginState");
 
-describe("WithRedirectIfLoggedOut", () => {
+describe("WithRedirectIfLoggedIn", () => {
   afterEach(() => {
     jest.clearAllMocks();
   });
@@ -20,20 +21,19 @@ describe("WithRedirectIfLoggedOut", () => {
     return <div>Dummy Component</div>;
   };
 
-  const WrappedComponent = WithRedirectIfLoggedOut(DummyComponent);
+  const WrappedComponent = WithRedirectIfLoggedIn(DummyComponent);
 
-  it("isLoginがfalseの時リダイレクトされる", () => {
-    (useGetLoginState as jest.Mock).mockReturnValue({ isLogin: false, userData: null });
-    render(<WrappedComponent />);
-    expect(replaceMock).toHaveBeenCalledWith("/auth/sign_in");
-  });
-
-  it("isLoginがtrueの時コンポーネントがレンダリングされる", () => {
+  it("isLoginがtrueの時リダイレクトされる", () => {
     (useGetLoginState as jest.Mock).mockReturnValue({
       isLogin: true,
       userData: validUserData
     });
+    render(<WrappedComponent />);
+    expect(replaceMock).toHaveBeenCalledWith("/");
+  });
 
+  it("isLoginがfalseの時コンポーネントがレンダリングされる", () => {
+    (useGetLoginState as jest.Mock).mockReturnValue({ isLogin: false, userData: null });
     const { getByText } = render(<WrappedComponent />);
     expect(getByText("Dummy Component")).toBeInTheDocument();
     expect(replaceMock).not.toHaveBeenCalled();
