+++ frontend/app/app/features/auth/accessControl/__tests__/WithRedirectIfLoggedOut.test.tsx
@@ -3,12 +3,13 @@ import { render } from "@testing-library/react";
 import { validUserData } from "@/__tests__/fixtures/auth/validUserData";
 import { mockNavigation, replaceMock } from "@/__tests__/mocks/mockNavigation";
 import { useGetLoginState } from "@/app/hooks/recoil/loginState/useGetLoginState";
-import WithRedirectIfLoggedIn from "@/app/components/HOCs/WithRedirectIfLoggedIn";
+
+import WithRedirectIfLoggedOut from "../WithRedirectIfLoggedOut";
 
 jest.mock("next/navigation", () => mockNavigation);
 jest.mock("@/app/hooks/recoil/loginState/useGetLoginState");
 
-describe("WithRedirectIfLoggedIn", () => {
+describe("WithRedirectIfLoggedOut", () => {
   afterEach(() => {
     jest.clearAllMocks();
   });
@@ -20,19 +21,20 @@ describe("WithRedirectIfLoggedIn", () => {
     return <div>Dummy Component</div>;
   };
 
-  const WrappedComponent = WithRedirectIfLoggedIn(DummyComponent);
+  const WrappedComponent = WithRedirectIfLoggedOut(DummyComponent);
 
-  it("isLoginがtrueの時リダイレクトされる", () => {
+  it("isLoginがfalseの時リダイレクトされる", () => {
+    (useGetLoginState as jest.Mock).mockReturnValue({ isLogin: false, userData: null });
+    render(<WrappedComponent />);
+    expect(replaceMock).toHaveBeenCalledWith("/auth/sign_in");
+  });
+
+  it("isLoginがtrueの時コンポーネントがレンダリングされる", () => {
     (useGetLoginState as jest.Mock).mockReturnValue({
       isLogin: true,
       userData: validUserData
     });
-    render(<WrappedComponent />);
-    expect(replaceMock).toHaveBeenCalledWith("/");
-  });
 
-  it("isLoginがfalseの時コンポーネントがレンダリングされる", () => {
-    (useGetLoginState as jest.Mock).mockReturnValue({ isLogin: false, userData: null });
     const { getByText } = render(<WrappedComponent />);
     expect(getByText("Dummy Component")).toBeInTheDocument();
     expect(replaceMock).not.toHaveBeenCalled();
